+++ enterprise_data/admin_analytics/database/queries/fact_enrollment_admin_dash.py
@@ -98,7 +98,7 @@ class FactEnrollmentAdminDashQueries:
         """
 
     @staticmethod
-    def get_top_courses_by_enrollments_query(record_count=20):
+    def get_top_courses_by_enrollments_query(record_count=10):
         """
         Get the query to fetch the enrollment count by courses.
 
@@ -108,15 +108,32 @@ class FactEnrollmentAdminDashQueries:
             record_count (int): Number of records to fetch.
         """
         return f"""
-            SELECT course_key, course_title , enroll_type, count(course_key) as enrollment_count
-            FROM fact_enrollment_admin_dash
-            WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
+            WITH filtered_data AS (
+                SELECT *
+                FROM fact_enrollment_admin_dash
+                WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
                 enterprise_enrollment_date BETWEEN %(start_date)s AND %(end_date)s
-            GROUP BY course_key, course_title, enroll_type ORDER BY enrollment_count DESC LIMIT {record_count};
+            ),
+            top_10_courses AS (
+                SELECT course_key
+                FROM filtered_data
+                GROUP BY course_key
+                ORDER BY COUNT(*) DESC
+                LIMIT {record_count}
+            )
+
+            SELECT
+                d.course_key,
+                d.enroll_type,
+                COUNT(*) AS enrollment_count
+            FROM filtered_data d
+            JOIN top_10_courses tc
+                ON d.course_key = tc.course_key
+            GROUP BY d.course_key, d.enroll_type;
         """
 
     @staticmethod
-    def get_top_subjects_by_enrollments_query(record_count=20):
+    def get_top_subjects_by_enrollments_query(record_count=10):
         """
         Get the query to fetch the enrollment count by subjects.
 
@@ -126,11 +143,27 @@ class FactEnrollmentAdminDashQueries:
             record_count (int): Number of records to fetch.
         """
         return f"""
-            SELECT course_subject, enroll_type, count(course_subject) enrollment_count
-            FROM fact_enrollment_admin_dash
-            WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
+            WITH filtered_data AS (
+                SELECT *
+                FROM fact_enrollment_admin_dash
+                WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
                 enterprise_enrollment_date BETWEEN %(start_date)s AND %(end_date)s
-            GROUP BY course_subject, enroll_type ORDER BY enrollment_count DESC LIMIT {record_count};
+            ),
+            top_10_subjects AS (
+                SELECT course_subject
+                FROM filtered_data
+                GROUP BY course_subject
+                ORDER BY COUNT(*) DESC
+                LIMIT {record_count}
+            )
+            SELECT
+                d.course_subject,
+                d.enroll_type,
+                COUNT(*) AS enrollment_count
+            FROM filtered_data d
+            JOIN top_10_subjects ts
+                ON d.course_subject = ts.course_subject
+            GROUP BY d.course_subject, d.enroll_type;
         """
 
     @staticmethod
@@ -162,7 +195,7 @@ class FactEnrollmentAdminDashQueries:
         """
 
     @staticmethod
-    def get_top_courses_by_completions_query(record_count=20):
+    def get_top_courses_by_completions_query(record_count=10):
         """
         Get the query to fetch the completion count by courses.
 
@@ -175,17 +208,40 @@ class FactEnrollmentAdminDashQueries:
             (str): Query to fetch the enrollment count by courses for the top courses by enrollment count.
         """
         return f"""
-            SELECT course_key, course_title, enroll_type, count(course_key) as completion_count
-            FROM fact_enrollment_admin_dash
-            WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
-                has_passed=1 AND
+            WITH filtered_data AS (
+                SELECT
+                    course_key,
+                    course_title,
+                    enroll_type,
+                    passed_date
+                FROM fact_enrollment_admin_dash
+                WHERE has_passed = 1 AND
+                enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
                 passed_date BETWEEN %(start_date)s AND %(end_date)s
-            GROUP BY course_key, course_title, enroll_type
-            ORDER BY completion_count DESC LIMIT {record_count};
+            ),
+            top_10_courses AS (
+                SELECT
+                    course_key,
+                    COUNT(*) AS total_completion_count
+                FROM filtered_data
+                GROUP BY course_key
+                ORDER BY total_completion_count DESC
+                LIMIT {record_count}
+            )
+            SELECT
+                d.course_key,
+                d.course_title,
+                d.enroll_type,
+                COUNT(*) AS completion_count
+            FROM filtered_data d
+            JOIN top_10_courses tc
+                ON d.course_key = tc.course_key
+            GROUP BY d.course_key, d.course_title, d.enroll_type
+            ORDER BY total_completion_count DESC;
         """
 
     @staticmethod
-    def get_top_subjects_by_completions_query(record_count=20):
+    def get_top_subjects_by_completions_query(record_count=10):
         """
         Get the query to fetch the completion count by subjects.
 
@@ -198,13 +254,34 @@ class FactEnrollmentAdminDashQueries:
             (str): Query to fetch the completion count by subjects for the top subjects by completion count.
         """
         return f"""
-            SELECT course_subject, enroll_type, count(course_subject) as completion_count
-            FROM fact_enrollment_admin_dash
-            WHERE enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
-                has_passed=1 AND
+            WITH filtered_data AS (
+                SELECT
+                    course_subject,
+                    enroll_type,
+                    passed_date
+                FROM fact_enrollment_admin_dash
+                WHERE has_passed = 1 AND
+                enterprise_customer_uuid=%(enterprise_customer_uuid)s AND
                 passed_date BETWEEN %(start_date)s AND %(end_date)s
-            GROUP BY course_subject, enroll_type
-            ORDER BY completion_count DESC LIMIT {record_count};
+            ),
+            top_10_subjects AS (
+                SELECT
+                    course_subject,
+                    COUNT(*) AS total_completion_count
+                FROM filtered_data
+                GROUP BY course_subject
+                ORDER BY total_completion_count DESC
+                LIMIT {record_count}
+            )
+            SELECT
+                d.course_subject,
+                d.enroll_type,
+                COUNT(*) AS completion_count
+            FROM filtered_data d
+            JOIN top_10_subjects ts
+                ON d.course_subject = ts.course_subject
+            GROUP BY d.course_subject, d.enroll_type
+            ORDER BY total_completion_count DESC;
         """
 
     @staticmethod