+++ enterprise_data/admin_analytics/database/tables/fact_engagement_admin_dash.py
@@ -5,10 +5,13 @@ from datetime import date
 from uuid import UUID
 
 from enterprise_data.cache.decorators import cache_it
+from enterprise_data.utils import find_first
 
 from ..queries import FactEngagementAdminDashQueries
 from ..utils import run_query
 from .base import BaseTable
+
+NULL_EMAIL_TEXT = 'learners who have not shared consent'
 
 
 class FactEngagementAdminDashTable(BaseTable):
@@ -168,7 +171,14 @@ class FactEngagementAdminDashTable(BaseTable):
 
     @cache_it()
     def _get_engagement_data_for_leaderboard(
-            self, enterprise_customer_uuid: UUID, start_date: date, end_date: date, limit: int, offset: int
+            self,
+            enterprise_customer_uuid: UUID,
+            start_date: date,
+            end_date: date,
+            limit: int,
+            offset: int,
+            include_null_email: bool,
+
     ):
         """
         Get the engagement data for leaderboard.
@@ -182,11 +192,12 @@ class FactEngagementAdminDashTable(BaseTable):
             end_date (date): The end date.
             limit (int): The maximum number of records to return.
             offset (int): The number of records to skip.
+            include_null_email (bool): If True, only fetch data for NULL emails.
 
         Returns:
             list[dict]: The engagement data for leaderboard.
         """
-        return run_query(
+        engagements = run_query(
             query=self.queries.get_engagement_data_for_leaderboard_query(),
             params={
                 'enterprise_customer_uuid': enterprise_customer_uuid,
@@ -198,9 +209,27 @@ class FactEngagementAdminDashTable(BaseTable):
             as_dict=True,
         )
 
+        if include_null_email:
+            engagement_for_null_email = run_query(
+                query=self.queries.get_engagement_data_for_leaderboard_null_email_only_query(),
+                params={
+                    'enterprise_customer_uuid': enterprise_customer_uuid,
+                    'start_date': start_date,
+                    'end_date': end_date,
+                },
+                as_dict=True,
+            )
+            engagements += engagement_for_null_email
+        return engagements
+
     @cache_it()
     def _get_completion_data_for_leaderboard_query(
-            self, enterprise_customer_uuid: UUID, start_date: date, end_date: date, email_list: list
+            self,
+            enterprise_customer_uuid: UUID,
+            start_date: date,
+            end_date: date,
+            email_list: list,
+            include_null_email: bool,
     ):
         """
         Get the completion data for leaderboard.
@@ -213,11 +242,13 @@ class FactEngagementAdminDashTable(BaseTable):
             start_date (date): The start date.
             end_date (date): The end date.
             email_list (list<str>): List of emails of the enterprise learners.
+            include_null_email (bool): If True, only fetch data for NULL emails.
 
         Returns:
             list[dict]: The engagement data for leaderboard.
         """
-        return run_query(
+
+        completions = run_query(
             query=self.queries.get_completion_data_for_leaderboard_query(email_list),
             params={
                 'enterprise_customer_uuid': enterprise_customer_uuid,
@@ -227,8 +258,28 @@ class FactEngagementAdminDashTable(BaseTable):
             as_dict=True,
         )
 
+        if include_null_email:
+            completions_for_null_email = run_query(
+                query=self.queries.get_completion_data_for_leaderboard_null_email_only_query(),
+                params={
+                    'enterprise_customer_uuid': enterprise_customer_uuid,
+                    'start_date': start_date,
+                    'end_date': end_date,
+                },
+                as_dict=True,
+            )
+            completions += completions_for_null_email
+
+        return completions
+
     def get_all_leaderboard_data(
-            self, enterprise_customer_uuid: UUID, start_date: date, end_date: date, limit: int, offset: int
+            self,
+            enterprise_customer_uuid: UUID,
+            start_date: date,
+            end_date: date,
+            limit: int,
+            offset: int,
+            total_count: int,
     ):
         """
         Get the leaderboard data for the given enterprise customer.
@@ -239,31 +290,48 @@ class FactEngagementAdminDashTable(BaseTable):
             end_date (date): The end date.
             limit (int): The maximum number of records to return.
             offset (int): The number of records to skip.
+            total_count (int): The total number of records.
 
         Returns:
             list[dict]: The leaderboard data.
         """
+        include_null_email = False
+        # If this is the last or only page, we need to include NULL emails record.
+        if total_count <= offset + limit:
+            include_null_email = True
+
         engagement_data = self._get_engagement_data_for_leaderboard(
             enterprise_customer_uuid=enterprise_customer_uuid,
             start_date=start_date,
             end_date=end_date,
             limit=limit,
             offset=offset,
+            include_null_email=include_null_email,
         )
         # If there is no data, no need to proceed.
         if not engagement_data:
             return []
 
-        engagement_data_dict = {engagement['email']: engagement for engagement in engagement_data}
+        engagement_data_dict = {
+            engagement['email']: engagement for engagement in engagement_data if engagement['email']
+        }
         completion_data = self._get_completion_data_for_leaderboard_query(
             enterprise_customer_uuid=enterprise_customer_uuid,
             start_date=start_date,
             end_date=end_date,
             email_list=list(engagement_data_dict.keys()),
+            include_null_email=include_null_email,
         )
         for completion in completion_data:
             email = completion['email']
             engagement_data_dict[email]['course_completion_count'] = completion['course_completion_count']
+
+        if include_null_email:
+            engagement_data_dict['None'] = find_first(engagement_data, lambda x: x['email'] is None) or {}
+            completion = find_first(completion_data, lambda x: x['email'] is None) or \
+                {'course_completion_count': ''}
+            engagement_data_dict['None']['course_completion_count'] = completion['course_completion_count']
+            engagement_data_dict['None']['email'] = NULL_EMAIL_TEXT
 
         return list(engagement_data_dict.values())