+++ enterprise_data/api/v1/views/analytics_completions.py
@@ -75,7 +75,7 @@ class AdvanceAnalyticsCompletionsView(AnalyticsPaginationMixin, ViewSet):
         )
 
         if response_type == ResponseType.CSV.value:
-            filename = f"""individual_completions, {start_date} - {end_date}.csv"""
+            filename = f"""Individual Completions, {start_date} - {end_date}.csv"""
 
             return StreamingHttpResponse(
                 IndividualCompletionsCSVRenderer().render(self._stream_serialized_data(