+++ CHANGELOG.rst
@@ -16,14 +16,32 @@ Unreleased
 
 =========================
 
+[10.5.0] - 2024-11-14
+---------------------
+  * Fix CSV file names
+  * Fix ordering of skills charts data
+
+[10.4.0] - 2024-11-14
+---------------------
+  * Updated text for null emails record of leaderboard.
+
+[10.3.0] - 2024-11-13
+---------------------
+  * Re-write top 10 charts queries for Enrollments, Engagements and Completions
+
+[10.2.0] - 2024-11-12
+---------------------
+  * Fixed null email issue for leaderboard.
+
+
 [10.1.0] - 2024-10-29
 ---------------------
   * Added management command to pre-warm analytics data.
 
 [10.0.1] - 2024-10-25
 ---------------------
-  * Same as ``10.0.0`` 
-  * Bumping the version so a new tag can be created in the GitHub 
+  * Same as ``10.0.0``
+  * Bumping the version so a new tag can be created in the GitHub
 
 [10.0.0] - 2024-10-25
 ---------------------